+++ src/api/aptos.ts
@@ -12,7 +12,6 @@ import { General } from "./general";
 import { ANS } from "./ans";
 import { Staking } from "./staking";
 import { Transaction } from "./transaction";
-import { TransactionSubmission } from "./transactionSubmission";
 
 /**
  * This class is the main entry point into Aptos's
@@ -44,8 +43,6 @@ export class Aptos {
 
   readonly transaction: Transaction;
 
-  readonly transactionSubmission: TransactionSubmission;
-
   constructor(settings?: AptosConfig) {
     this.config = new AptosConfig(settings);
     this.account = new Account(this.config);
@@ -58,12 +55,13 @@ export class Aptos {
     this.general = new General(this.config);
     this.staking = new Staking(this.config);
     this.transaction = new Transaction(this.config);
-    this.transactionSubmission = new TransactionSubmission(this.config);
   }
 }
 
+// extends Aptos interface so all the methods and properties from the other classes will be recognized by typescript.
 export interface Aptos
   extends Account,
+    ANS,
     Coin,
     DigitalAsset,
     Event,
@@ -71,9 +69,7 @@ export interface Aptos
     FungibleAsset,
     General,
     Staking,
-    Transaction,
-    ANS,
-    TransactionSubmission {}
+    Transaction {}
 
 /**
 In TypeScript, we can’t inherit or extend from more than one class,
@@ -106,4 +102,3 @@ applyMixin(Aptos, FungibleAsset, "fungibleAsset");
 applyMixin(Aptos, General, "general");
 applyMixin(Aptos, Staking, "staking");
 applyMixin(Aptos, Transaction, "transaction");
-applyMixin(Aptos, TransactionSubmission, "transactionSubmission");