+++ tests/e2e/api/transaction.test.ts
@@ -20,18 +20,18 @@ describe("transaction api", () => {
     const senderAccount = Account.generate();
     await aptos.fundAccount({ accountAddress: senderAccount.accountAddress, amount: FUND_AMOUNT });
     const bob = Account.generate();
-    const rawTxn = await aptos.build.transaction({
+    const rawTxn = await aptos.transaction.build.simple({
       sender: senderAccount.accountAddress,
       data: {
         function: "0x1::aptos_account::transfer",
         functionArguments: [bob.accountAddress, new U64(10)],
       },
     });
-    const authenticator = aptos.sign.transaction({
+    const authenticator = aptos.transaction.sign({
       signer: senderAccount,
       transaction: rawTxn,
     });
-    const response = await aptos.submit.transaction({
+    const response = await aptos.transaction.submit.simple({
       transaction: rawTxn,
       senderAuthenticator: authenticator,
     });
@@ -45,18 +45,18 @@ describe("transaction api", () => {
       const senderAccount = Account.generate();
       await aptos.fundAccount({ accountAddress: senderAccount.accountAddress, amount: FUND_AMOUNT });
       const bob = Account.generate();
-      const rawTxn = await aptos.build.transaction({
+      const rawTxn = await aptos.transaction.build.simple({
         sender: senderAccount.accountAddress,
         data: {
           function: "0x1::aptos_account::transfer",
           functionArguments: [bob.accountAddress, new U64(10)],
         },
       });
-      const authenticator = aptos.sign.transaction({
+      const authenticator = aptos.transaction.sign({
         signer: senderAccount,
         transaction: rawTxn,
       });
-      const response = await aptos.submit.transaction({
+      const response = await aptos.transaction.submit.simple({
         transaction: rawTxn,
         senderAuthenticator: authenticator,
       });