+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Add release automation, so version updates can be made with simply `pnpm update-version`
 - Rename custom request header to `aptos-typescript-sdk`
 - [`Breaking`] Rename `token` to `digitalAsset` and add digital asset built in transaction generation functions
+- [`Breaking`] change transaction submission builder flow namespace to be under a `transaction` namespace
 
 ## 0.0.8 (2023-11-29)